Implement parameter expansion in token list
The `expand_parameter` function expands parameters in tokens of type `WORD`
within a given token list (`t_tokenlist`). It iterates through the list,
invoking `expand_parameters_in_a_token` for each `WORD` token to handle the
expansion process. This involves splitting the token into subtokens, expanding
parameters, and reinserting the expanded fields back into the token list.
The function returns `0` on success and `-1` on failure.

+++ sources/parser/parsetree.c
@@ -6,7 +6,7 @@
 /*   By: seonseo <seonseo@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/18 16:58:24 by seonseo           #+#    #+#             */
-/*   Updated: 2024/06/01 20:32:27 by seonseo          ###   ########.fr       */
+/*   Updated: 2024/06/18 10:33:51 by seonseo          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -60,7 +60,7 @@ void	free_ast_node(t_ast_node *node)
 
 void	clear_ast(t_ast_node *node)
 {
-	int	i;
+	size_t	i;
 
 	if (node == NULL)
 		return ;
